Split out a session construct
